refactor: merge tests, make test name explicit to behaviour
